#OSIS-710 Refactoring suite à la review
